Fixes
Please remember timestamps or migration will not be applied.
In order to attack foreign key constraint, the target table must exist.
Work around is to make the table then add the constraint.  Dropping the
target table will remove the fk constraint so for rollback it's fine to
just dropt he table